# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Teamwork API
  description: An API for an internal social network for employees of an organization.
  version: 1.0.0
servers:
  - url: https://team-worker.herokuapp.com/api/v1
    description: Main (production) server
  - url: http://localhost:3000/api/v1
    description: Internal development server for local testing
...
paths:
...
  /gifs:
    post:
      tags:
        - gif
      operationId: createGif
      summary: Create a new gif
      description: Create a new gif
      requestBody:
        description: Create a new gif
        content:
          application/json:
            schema: 
              type: object
              required:
                - title
                - image
              properties:
                title:
                  type: string
                  description: Title of gif
                  example: How to become a 10x software developer
                image:
                  type: string
                  description: image/gif content
                  example: /https://imgbb.com/
      responses:
        '201':
          description: A successful operation
          content: 
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: successful
                    example: success
                  data:
                    type: object
                    description: Data for new gif created
                    properties:
                      message:
                        type: string
                        description: Message for gif created
                        example: GIF successfully posted
                      userId:
                        type: integer
                        description: user id
                        example: 10
                      gifId:
                        type: integer
                        description: Gif id
                        example: 10
                      createdOn: 
                        type: string
                        description: Created date of a new gif
                        pattern: /([0-9]{4})-(?:[0-9]{2})-([0-9]{2})/
                        example: "2019-05-17"
                      title: 
                        type: string
                        description: Title of gif
                        example: How to become a 10x software developer
                      published:
                        type: boolean
                        description: Published or not
                        example: True
                      imageUrl:
                        type: string
                        description: image/gif content
                        example: /https://imgbb.com/
      security:
        - access_token: []
  /gifs/{gifId}:
    get:
      tags:
        - gif
      operationId: getGif
      summary: Get a gif
      description: Get a gif with a gif id
      parameters:
      - in : path
        name : gifId
        schema:
          type: integer
        required: true
        description : Numeric ID of the gif to get
      responses:
        '200':
          description: A successfully Operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: success
                    example: success
                  data:
                    type: object
                    description: Gif was gotten successfuly
                    properties:
                      gifId:
                        type: integer
                        description: Gif id
                        example: 10
                      createdOn: 
                        type: string
                        description: Created date of a new gif
                        pattern: /([0-9]{4})-(?:[0-9]{2})-([0-9]{2})/
                        example: "2019-05-17"
                      title: 
                        type: string
                        description: Title of gif
                        example: How to become a 10x software developer
                      imageUrl: 
                        type: string
                        description: url for image/gif
                        example: /https://imgbb.com/
                      published:
                        type: boolean
                        description: Published or not
                        example: True
                      comments:
                        type: array
                        items:
                          properties:
                            commentId:
                              type: integer
                              description: Comment Id
                              example: 10
                            comment:
                              type: string
                              description: New comment to be created
                              example: This post articulates all there is to be a 10x software developer
                            authorId:
                              type: integer
                              description: Comment Id
                              example: 10
      security:
          - access_token: []
    delete:
      tags:
        - gif
      operationId: deleteGif
      summary: delete a gif
      description: Delete a gif with a gif id
      parameters:
      - in : path
        name : gifId
        schema:
          type: integer
        required: true
        description : Numeric ID of the gif to get
      responses:
        '200':
          description: A successfully Operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: success
                    example: success
                  data:
                    type: object
                    description: GIF image deleted successfuly
                    properties:
                      message:
                        type: string
                        description: Message to indicate successful delete
                        example: GIF image successfully deleted 
  /gifs/{gifId}/comment:
    post:
      tags:
        - gif
      operationId: Create a comment for a gif
      summary: Create a comment for a gif
      description: Create a comment for a gif
      parameters:
      - in : path
        name : gifId
        schema:
          type: integer
        required: true
        description : Numeric ID of the gif to get
      requestBody:
        description: Create a new comment for a gif
        content:
          application/json:
            schema:
              type: object
              properties:
                userId: 
                  type: integer
                  description: user id
                  example: 10
                comment:
                  type: string
                  description: New comment to be created
                  example: This post articulates all there is to be a 10x software developer
      responses:
        '201':
          description: Successful creation of comment for gif
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: successful creation of comment
                    example: success
                  data:
                    type: object
                    description: Data for new comment created on gif
                    properties:
                      message:
                        type: string
                        description: Message for comment created on gif
                        example: Comment successfully created
                      createdOn: 
                        type: string
                        description: Created date of a new gif
                        pattern: /([0-9]{4})-(?:[0-9]{2})-([0-9]{2})/
                        example: "2019-05-17"
                      gifTitle: 
                        type: string
                        description: Title of gif
                        example: How to become a 10x software developer
                      image: 
                        type: string
                        description: url for image/gif
                        example: /https://imgbb.com/
                      comment:
                        type: string
                        description: New comment to be created
                        example: This post articulates all there is to be a 10x software developer
      security:
        - access_token: []
  /gifs/{gifId}/likes:
    post:
      tags:
        - gif
      operationId: Record a like for a gif
      summary: Record a like for a gif
      description: Record a like for a gif
      parameters:
      - in : path
        name : gifId
        schema:
          type: integer
        required: true
        description : Numeric ID of the gif to get
      requestBody:
        description: Create a new like for a gif
        content:
          application/json:
            schema:
              type: object
              properties:
                userId: 
                  type: integer
                  description: user id
                  example: 10
      responses:
        '201':
          description: Successful recording a like for a gif
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: successful liking a gif
                    example: success
                  data:
                    type: object
                    description: Data for new like created on gif
                    properties:
                      message:
                        type: string
                        description: Message for like created on gif
                        example: Gif  successfully liked
                      userId: 
                        type: integer
                        description: user id
                        example: 10
                      gifId: 
                        type: integer
                        description: gif id
                        example: 10
      security:
        - access_token: []
...